+++ src/backend/ci/core/auth/biz-auth-blueking/src/main/kotlin/com/tencent/devops/auth/AuthConfiguration.kt
@@ -31,37 +31,21 @@ import com.tencent.devops.auth.service.SimpleAuthPermissionProjectService
 import com.tencent.devops.auth.service.SimpleAuthPermissionService
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.tencent.bk.sdk.iam.config.IamConfiguration
-import com.tencent.bk.sdk.iam.helper.AuthHelper
 import com.tencent.bk.sdk.iam.service.ActionService
-import com.tencent.bk.sdk.iam.service.PolicyService
-import com.tencent.bk.sdk.iam.service.ResourceService
-import com.tencent.bk.sdk.iam.service.SystemService
 import com.tencent.bk.sdk.iam.service.impl.ApigwHttpClientServiceImpl
 import com.tencent.bk.sdk.iam.service.impl.ManagerServiceImpl
 import com.tencent.devops.auth.dao.ActionDao
 import com.tencent.devops.auth.service.AuthDeptServiceImpl
-import com.tencent.devops.auth.service.AuthGroupService
-import com.tencent.devops.auth.service.BkPermissionProjectService
-import com.tencent.devops.auth.service.BkPermissionService
-import com.tencent.devops.auth.service.DeptService
 import com.tencent.devops.auth.service.action.BkResourceService
-import com.tencent.devops.auth.service.action.impl.IamBkActionServiceImpl
-import com.tencent.devops.auth.service.action.impl.IamBkResourceServiceImpl
 import com.tencent.devops.auth.service.action.impl.SimpleBkActionServiceImpl
 import com.tencent.devops.auth.service.action.impl.SimpleBkResourceServiceImpl
-import com.tencent.devops.auth.service.iam.IamCacheService
 import com.tencent.devops.auth.service.iam.PermissionProjectService
-import com.tencent.devops.auth.service.iam.PermissionRoleMemberService
-import com.tencent.devops.auth.service.iam.PermissionRoleService
 import com.tencent.devops.auth.service.iam.PermissionService
 import com.tencent.devops.auth.service.stream.GithubStreamPermissionServiceImpl
 import com.tencent.devops.auth.service.stream.GitlabStreamPermissionServiceImpl
 import com.tencent.devops.auth.service.stream.StreamPermissionProjectServiceImpl
 import com.tencent.devops.auth.service.stream.StreamPermissionServiceImpl
-import com.tencent.devops.common.auth.api.AuthProjectApi
-import com.tencent.devops.common.auth.code.BluekingV3ProjectAuthServiceCode
 import com.tencent.devops.common.auth.service.IamEsbService
-import com.tencent.devops.common.client.Client
 import com.tencent.devops.common.redis.RedisOperation
 import org.jooq.DSLContext
 import org.springframework.beans.factory.annotation.Value
@@ -122,43 +106,6 @@ class AuthConfiguration {
     ) = AuthDeptServiceImpl(redisOperation, objectMapper)
 
     @Bean
-    @ConditionalOnProperty(prefix = "auth", name = ["idProvider"], havingValue = "bk_login_v3")
-    fun v3permissionService(
-        authHelper: AuthHelper,
-        policyService: PolicyService,
-        iamConfiguration: IamConfiguration,
-        iamCacheService: IamCacheService
-    ) = BkPermissionService(authHelper, policyService, iamConfiguration, iamCacheService)
-
-    @Bean
-    @ConditionalOnProperty(prefix = "auth", name = ["idProvider"], havingValue = "bk_login_v3")
-    fun v3permissionProjectService(
-        permissionRoleService: PermissionRoleService,
-        permissionRoleMemberService: PermissionRoleMemberService,
-        authHelper: AuthHelper,
-        policyService: PolicyService,
-        client: Client,
-        iamConfiguration: IamConfiguration,
-        deptService: DeptService,
-        groupService: AuthGroupService,
-        iamCacheService: IamCacheService,
-        authProjectApi: AuthProjectApi,
-        projectAuthServiceCode: BluekingV3ProjectAuthServiceCode
-    ) = BkPermissionProjectService(
-        permissionRoleService = permissionRoleService,
-        permissionRoleMemberService = permissionRoleMemberService,
-        authHelper = authHelper,
-        policyService = policyService,
-        client = client,
-        iamConfiguration = iamConfiguration,
-        deptService = deptService,
-        groupService = groupService,
-        iamCacheService = iamCacheService,
-        authProjectApi = authProjectApi,
-        projectAuthServiceCode = projectAuthServiceCode
-    )
-
-    @Bean
     @ConditionalOnProperty(prefix = "auth", name = ["idProvider"], havingValue = "github")
     fun githubStreamPermissionService() = GithubStreamPermissionServiceImpl()
 
@@ -189,31 +136,4 @@ class AuthConfiguration {
     @Bean
     @ConditionalOnMissingBean(BkResourceService::class)
     fun simpleResourceService() = SimpleBkResourceServiceImpl()
-
-    @Bean
-    @ConditionalOnProperty(prefix = "auth", name = ["idProvider"], havingValue = "bk_login_v3")
-    fun iamActionService(
-        dslContext: DSLContext,
-        actionDao: ActionDao,
-        resourceService: BkResourceService,
-        iamConfiguration: IamConfiguration,
-        systemService: SystemService,
-        iamActionService: ActionService,
-        iamResourceService: ResourceService
-    ) = IamBkActionServiceImpl(
-        dslContext = dslContext,
-        actionDao = actionDao,
-        resourceService = resourceService,
-        iamConfiguration = iamConfiguration,
-        systemService = systemService,
-        iamActionService = iamActionService,
-        iamResourceService = iamResourceService
-    )
-
-    @Bean
-    @ConditionalOnProperty(prefix = "auth", name = ["idProvider"], havingValue = "bk_login_v3")
-    fun iamResourceService(
-        iamConfiguration: IamConfiguration,
-        resourceService: ResourceService
-    ) = IamBkResourceServiceImpl(iamConfiguration, resourceService)
 }
